Split results form components into their own files.
